Added the 'copy.odin32.src' target to 'buildFragment.xml'
It copies the Platform Implementation sources to a work-directory.
This includes the C/C++ sources for the JNI layer, which produce the
platform DLLs, as well as the Java sources, which produce the platform
class files. This process is known as 'gathering' the sources.

It is a copy of the 'copy.win32.src' target so it can be modified
in the future without interfering with the Win32 build-stuff.

Specific parts of the Java sources are initially 64-bit and use the
64-bit 'long' type, where 32-bit 'int' is required on platforms such
as Win32 and OS/2. This is handled during the 'gathering' where source
lines are translated based on specific annotations.

This means that after the gathering of the sources, the versions in
the work-directory have been translated to the bit-width of the target
platform and should not be copied back to their origins as a whole.
This only applies to the Java sources and not the C/C++ sources.

The 'copy.odin32.src' target expects to have some variables pre-setup,
especially the location of the work-directory where the translated
sources are copied to and where the actual building takes place.
Below is an excerpt of a GNUmakefile target 'gather-sources':

# Properties required for the Ant Targets
ANT_TARGET_PROPS:=
ANT_TARGET_PROPS+=-Dplugindir=$(ORG_ECLIPSE_SWT)
ANT_TARGET_PROPS+=-Dtemp.folder=$(WORKDIR)
ANT_TARGET_PROPS+=-DbuildDir=$(WORKDIR)
ANT_TARGET_PROPS+=-DsrcDir=$(WORKDIR)/org
ANT_TARGET_PROPS+=-DdestDir=$(WORKDIR)/out
ANT_TARGET_PROPS+=-Dswt.ws=win32
ANT_TARGET_PROPS+=-Dswt.os=win32
ANT_TARGET_PROPS+=-Dswt.arch=x86

# Gather the platform implementation sources
gather-sources:
	@$(ANT) -f buildFragment.xml copy.odin32.src $(ANT_TARGET_PROPS)