Update Build Level
For the library (swt.dll and swt-awt.dll) this was the level distributed
with the Alpha-1 Demo available at:
https://github.com/bitwiseworks/eclipse.platform.swt/releases/tag/SWT_OS2-Alpha1

So the source-archive accompanied with the Alpha-1 Demo is slightly
behind. Next time having the Demo ready will be part of the Milestone
so that the accompanied sources are in sync with level the Demo is
running on.